+++ plot1.R
@@ -1,10 +1,15 @@
+#read remote file and unzip it
 
-url<- "https://d396qusza40orc.cloudfront.net/exdata%2Fdata%2FNEI_data.zip"
-download.file(url, "exdata_data_NEI_data.zip")
-unzip("exdata_data_NEI_data.zip", exdir = "data")
-file.remove("exdata_data_NEI_data.zip")
+if (!file.exists("data/summarySCC_PM25.rds")) 
+{
+  url<- "https://d396qusza40orc.cloudfront.net/exdata%2Fdata%2FNEI_data.zip"
+  download.file(url, "exdata_data_NEI_data.zip")
+  unzip("exdata_data_NEI_data.zip", exdir = "data")
+  file.remove("exdata_data_NEI_data.zip")
+}
 
-dir()
+
+dir("data")
 
 ## load data file 
 NEI <- readRDS("data/summarySCC_PM25.rds")
